requirements.txt
Django==5.1.4
djangorestframework==3.15.2
openai==1.53.0
google-generativeai==0.8.5
yt-dlp==2024.12.13
requests==2.32.3
python-dotenv==1.0.1
gunicorn==20.1.0
assemblyai==0.38.0
psycopg2-binary==2.9.10
ytmdl
